♻️refactor: react -> pages 하위 게시판별 PageComponent 정리 #127
