+++ devbox_client/src/pages/gatherMate/GatherMateDetail.css
@@ -1,5 +1,5 @@
 .Gather-detail {
-  background-color: white !;
+  background-color: white;
   border-radius: 8px;
   padding: 20px;
   box-shadow: 0 0 10px rgba(0,0,0,0.1);
