+++ devbox_client/src/pages/gatherMate/GatherMateEdit.jsx
@@ -1,11 +1,11 @@
 import { useState, useEffect } from "react";
 import { useParams, useNavigate } from "react-router-dom";
 import "react-quill/dist/quill.snow.css";
-import Button from "../components/Button";
-import WriteLong from "../components/WriteLong";
-import WriteShort from "../components/WriteShort";
-import QuillEditor from "../components/QuillEditor";
-import WriteSelect from "../components/WriteSelect";
+import Button from "../../components/Button";
+import WriteLong from "../../components/WriteLong";
+import WriteShort from "../../components/WriteShort";
+import QuillEditor from "../../components/QuillEditor";
+import WriteSelect from "../../components/WriteSelect";
 
 function GatherMateEdit() {
   const [intro, setIntro] = useState("");
